+++ Libplanet.Tests/Store/StoreTracker.cs
@@ -176,15 +176,15 @@ namespace Libplanet.Tests.Store
             return _store.IterateStateReferences(@namespace, address);
         }
 
-        public void StoreStateReference<T>(
+        public void StoreStateReference(
             string @namespace,
             IImmutableSet<Address> addresses,
-            Block<T> block)
-            where T : IAction, new()
+            HashDigest<SHA256> blockHash,
+            long blockIndex)
         {
             // FIXME: Log arguments properly (including @namespace).
-            _logs.Add((nameof(StoreStateReference), block.Hash, null));
-            _store.StoreStateReference(@namespace, addresses, block);
+            _logs.Add((nameof(StoreStateReference), blockHash, null));
+            _store.StoreStateReference(@namespace, addresses, blockHash, blockIndex);
         }
 
         public void ForkStateReferences<T>(