+++ Libplanet/Store/LiteDBStore.cs
@@ -432,10 +432,11 @@ namespace Libplanet.Store
         }
 
         /// <inheritdoc/>
-        public override void StoreStateReference<T>(
+        public override void StoreStateReference(
             string @namespace,
             IImmutableSet<Address> addresses,
-            Block<T> block)
+            HashDigest<SHA256> hash,
+            long index)
         {
             string collId = StateRefId(@namespace);
             LiteCollection<StateRefDoc> coll = _db.GetCollection<StateRefDoc>(collId);
@@ -443,8 +444,8 @@ namespace Libplanet.Store
                 addresses.Select(addr => new StateRefDoc
                 {
                     Address = addr,
-                    BlockIndex = block.Index,
-                    BlockHash = block.Hash,
+                    BlockIndex = index,
+                    BlockHash = hash,
                 })
             );
             coll.EnsureIndex("AddressString");